refactor: Refinement class を domain/refinement に移動
